+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 - fixed bug removing oncogenic fusions
 - GCGI-1217: Remove NORMAL_DEPTH variable  
 - GCGI-1194: Research Report PDF Footer says RUO 
-  
-Improvements for mini-Djerba:
+- GCGI-1261: Update plugin tester base class, so it is unaffected by changes in core HTML
+
+### Improvements for mini-Djerba
+
 - GCGI-1265: Automatically fill in report dates (preserving date of original report draft)
 - GCGI-1268: Use report ID for name of JSON output file
 - GCGI-1269: Default to user-friendly minimal error text